packages/ml-training/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "canopy-ml-training"
version = "1.0.0"
description = "Canopy Sight — YOLO11 + EfficientNet CNN training pipeline for wildlife & security detection"
requires-python = ">=3.10"
dependencies = [
    # Core ML
    "ultralytics>=8.3.0",        # YOLO11 (latest Ultralytics)
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "torchaudio>=2.3.0",
    "onnx>=1.16.0",
    "onnxruntime>=1.18.0",
    "onnxsim>=0.4.36",           # ONNX simplifier — reduces model size

    # Data / augmentation
    "albumentations>=1.4.0",
    "opencv-python-headless>=4.9.0",
    "Pillow>=10.3.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pycocotools>=2.0.7",

    # Dataset access
    "roboflow>=1.1.0",           # Roboflow public wildlife datasets
    "fiftyone>=0.23.0",          # Dataset visualization + curation

    # Experiment tracking
    "wandb>=0.17.0",             # Weights & Biases
    "tensorboard>=2.17.0",

    # Utilities
    "PyYAML>=6.0.1",
    "tqdm>=4.66.0",
    "scipy>=1.13.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.5.0",
    "rich>=13.7.0",              # Beautiful terminal output
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "black>=24.4.0",
    "ruff>=0.4.0",
]
gpu = [
    # Install CUDA-enabled torch separately via:
    # pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
]

[project.scripts]
canopy-train-detector  = "src.train_detector:main"
canopy-train-behavior  = "src.train_behavior:main"
canopy-export          = "src.export_onnx:main"
canopy-prepare-data    = "scripts.prepare_data:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*", "configs*"]
